third_party/blink/web_tests/platform/linux-chrome/external/wpt/infrastructure/testdriver/actions/multiTouchPointsSimultaneousMove-expected.txt

This is a testharness.js-based test.
[FAIL PASS] TestDriver actions: two touch points with both moving
  assert_equals: Event pointermove pointerId 3 property pointerId expected 2 but got 3
Harness: the test ran to completion.
